Checking times

+++ writes.cpp
@@ -28,7 +28,7 @@ int main(int argc, char** argv) {
     rocksdb::Status status = rocksdb::DB::Open(options, "/tmp/tree", &db);
     assert(status.ok());
 
-    for (long i = start; i < limit; i += 1) {
+    for (long long i = start; i < limit; i += 1) {
         std::string key = std::to_string(i);
         std::string value = std::to_string(i * 2);
         db->Put(rocksdb::WriteOptions(), key, value);
